New login API integration
