+++ src/components/auth/Login/LoginFormContent.jsx
@@ -1,21 +1,21 @@
-import StudentLoginForm from "./StudentLoginForm";
-import LoginImage from "../LoginImage";
+import LoginForm from "./LoginForm";
+import LoginImage from "../Login/LoginImage";
 
-const StudentLoginFormContent = () => {
+const LoginFormContent = () => {
   return (
     <div className="main-inner-wrapper flex items-stretch gap-2">
       <div className="image">
-        <LoginImage/>
+        <LoginImage />
       </div>
       <div className="form-wrapper w-96 px-[40px] py-[30px] bg-white rounded-md">
         <div className="header">
           <h2>Idea Bank</h2>
           <h3>Login</h3>
         </div>
-        <StudentLoginForm />
+        <LoginForm />
       </div>
     </div>
   );
 };
 
-export default StudentLoginFormContent;
+export default LoginFormContent;
