+++ src/components/auth/StudentRegistrationFormContent.jsx
@@ -1,11 +1,11 @@
 import StudentRegistrationForm from "./StudentRegistrationForm";
-import LoginImage from "../LoginImage";
+import LoginImage from "../auth/Login/LoginImage";
 
 const RegistrationFormContent = () => {
   return (
     <div className="main-inner-wrapper flex items-stretch gap-2">
       <div className="image">
-      <LoginImage/>
+        <LoginImage />
       </div>
       <div className="form-wrapper w-96 px-[40px] py-[30px] bg-white rounded-md">
         <div className="header">
