--- src/context/UserContext.jsx
@@ -1,56 +1,0 @@
-// import { createContext, useState } from "react";
-
-// export const UserContext = createContext();
-
-// export const UserProvider = ({ children }) => {
-//   const [currentUser, setCurrentUser] = useState(() => {
-//     const storedUser = localStorage.getItem("currentUser");
-//     return storedUser ? JSON.parse(storedUser) : null;
-//   });
-
-//   const loginUser = (user) => {
-//     setCurrentUser(user);
-//     localStorage.setItem("currentUser", JSON.stringify(user));
-//   };
-
-//   const logoutUser = () => {
-//     setCurrentUser(null);
-//     localStorage.removeItem("currentUser");
-//   };
-
-//   return (
-//     <UserContext.Provider value={{ currentUser, loginUser, logoutUser }}>
-//       {children}
-//     </UserContext.Provider>
-//   );
-// };
-
-
-import { createContext, useState } from "react";
-
-export const UserContext = createContext();
-
-export const UserProvider = ({ children }) => {
-  const [currentUser, setCurrentUser] = useState(() => {
-    const storedUser = localStorage.getItem("currentUser");
-    return storedUser ? JSON.parse(storedUser) : null;
-  });
-
-  const loginUser = (user) => {
-    setCurrentUser(user);
-    localStorage.setItem("currentUser", JSON.stringify(user));
-  };
-
-  const logoutUser = () => {
-    setCurrentUser(null);
-    localStorage.removeItem("currentUser");
-  };
-
-  const authToken = currentUser?.token || null;
-
-  return (
-    <UserContext.Provider value={{ currentUser, loginUser, logoutUser, authToken }}>
-      {children}
-    </UserContext.Provider>
-  );
-};
